Confirm item deletion

+++ app/src/main/java/com/example/teqelmasr/displaySellerProducts/view/MyProductsAdapter.kt
@@ -1,9 +1,12 @@
 package com.example.teqelmasr.displaySellerProducts.view
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.teqelmasr.R
@@ -12,8 +15,8 @@ import com.example.teqelmasr.model.Product
 
 class MyProductsAdapter(private val context: Context, private val onBtnListener: OnBtnListener): RecyclerView.Adapter<MyProductsAdapter.ViewHolder>() {
 
-    private var productList: List<Product> = listOf()
-    private val TAG = "MyProducts Adapter"
+    private var productList: ArrayList<Product> = ArrayList()
+    private val TAG = "MyProductsAdapter"
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
         ViewHolder(MyProductsListItemBinding.inflate(LayoutInflater.from(parent.context)))
 
@@ -26,7 +29,21 @@ class MyProductsAdapter(private val context: Context, private val onBtnListener:
             price.text = "${currentItem.variants!![0]?.price.toString()} LE"
 
             deleteBtn.setOnClickListener {
-                onBtnListener.onDeleteClick(currentItem)
+                val builder = AlertDialog.Builder(context)
+                builder.setMessage(R.string.delete_message)
+                    .setPositiveButton(R.string.yes
+                    ) { dialog, _ ->
+                        onBtnListener.onDeleteClick(currentItem)
+                        productList.removeAt(position)
+                        notifyDataSetChanged()
+                        dialog.dismiss()
+                        Toast.makeText(context, R.string.item_deleted, Toast.LENGTH_SHORT).show()
+
+                    }
+                    .setNegativeButton(R.string.cancel){ dialog, _ ->
+                        dialog.dismiss()
+                    }
+                    .create().show()
             }
 
         }
@@ -35,8 +52,7 @@ class MyProductsAdapter(private val context: Context, private val onBtnListener:
 
     override fun getItemCount(): Int = productList.size
 
-    fun setData(products: List<Product>){
-        Log.i("TAG", "setData: ${products.size}")
+    fun setData(products: ArrayList<Product>){
         productList = products
         notifyDataSetChanged()
     }
